refactor: split route handlers into modules
